add pending animation
